Исправила замечания преподавателя

+++ project/include/utils.h
@@ -3,7 +3,7 @@
 
 #include <stddef.h>
 
-size_t timer_from(int from);
-int custom_pow(int base, int power);
+int timer_from(int from);
+float custom_pow(int base, int power);
 
 #endif  //  PROJECT_INCLUDE_UTILS_H_
